Showcasing the sketch:
Since the development focus is starting to shift on different additional tasks and tests, and the Jupyter Notebook's cells are getting more complicated, the sketch version is being presented, showcasing what this work is going to be looking like.

More commits related to this update are on the 'dev' branch under this commit.