Added the 'sigma-clipping' cell and the cells for the following plots:
A plot of the PSD of the kernel; a plot of the correlograms: direct one (with sigma mask) and the correlogram of the kernel; a plot of the MAP model; a plot of the residuals of this model.

More commits related to this update are on the 'dev' branch under this commit.